requirements.txt
black==24.3.0
hydra-core==1.3.2
isort==5.13.2
lava-dl @ https://github.com/lava-nc/lava-dl/releases/download/v0.5.0/lava_dl-0.5.0-py3-none-any.whl
lava-nc==0.9.0
matplotlib==3.8.2
mlflow==2.10.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
setproctitle==1.3.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.2
